move transposition pipeline to sequenes_lib
